Fix reversed switches false click (#189)
* GpioSwitch: fix update function debug print

The print says "low" and "high" irrespective of the "reverse" parameter.

This is opposite to reality when "reverse" is set.

* GpioSwitch: set pin state before direction

This is to avoid momentarily driving the pin in the incorrect level.

Brief phantom "presses" of a "reverse" GpioSwitch have been reported a
few times and this is the cause.

A good example is a GPO controlling a relay which controls a physical
button like a garage door opener remote wall control. On Pi boot, the
GpioOutput would set the mode to output, but pigpiod has the value 0
configured by default. This would result in a permanent press of the
button, which could i this example open a garage door.

Raspberry Pi's first few GPIO pins as pulled high on power on. This
means if you want to wire the Pi GPO pin directly to the relay, you must
absolutely must set the relay to active low, and use the N.O. relay
contacts to operate the button.

In this configuration (reverse and pulse), the desired "unpressed
switch" state must be programmed into PigpioD prior to setting the pin's
mode to output.

The "unpressed" depends on the "reverse" property.